crud completed

+++ src/main/java/com/example1ipc2/app/persistence/CrudDAO.java
@@ -10,8 +10,8 @@ import java.util.List;
 public abstract class CrudDAO<T> {
 
     public abstract T insert(T entity) throws SQLException;
-    public abstract T findById(Long id) throws SQLException;
+    public abstract T findById(Integer id) throws SQLException;
     public abstract List<T> findAll() throws SQLException;
     public abstract void update(T entity) throws SQLException;
-    public abstract void delete(Long id) throws SQLException;
+    public abstract void delete(Integer id) throws SQLException;
 }
